react-app: refactor portfolio api into panel component refs oursky/likedao#169
